% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markmap.R
\name{markmap}
\alias{markmap}
\title{Create a markmap widget}
\usage{
markmap(json, width = NULL, height = NULL, elementId = NULL,
  options = NULL)
}
\description{
This function creates a markmap widget using htmlwidgets.The widget can be rendered on HTML pages generated from R Markdown,Shiny,or other applications.
Options for markmap creation#'
}
\examples{
gtor.json<-system.file('examples/gtor.json',package = 'Rmarkmap')
markmap(gtor.json)
}

