% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markmap.R
\name{markmap}
\alias{markmap}
\title{Create a markmap widget}
\usage{
markmap(path, width = NULL, height = NULL, elementId = NULL,
  options = markmapOption())
}
\arguments{
\item{path}{the path of markdown file}

\item{width}{the width of the markmap}

\item{height}{the height of the markmap}

\item{options}{the markmap options}
}
\value{
A HTML widget object rendered from a given document.
}
\description{
This function creates a markmap widget using htmlwidgets.
The widget can be rendered on HTML pages generated from
R Markdown,Shiny,or other applications.
}
\examples{
md<-system.file('examples/test.md',package = 'Rmarkmap')
markmap(md)
}

